Remove email reading list and restyle email signup
