--- client/components/reading-list/main.scss
@@ -1,124 +1,0 @@
-.myft-email-signup,
-.myft-reading-list {
-	display: none;
-
-	&__loaded {
-		display: block;
-		margin-bottom: 20px;
-	}
-
-
-	.more-on__content {
-		overflow: hidden;
-		background: oColorsGetPaletteColor('warm-2');
-	}
-
-	.more-on__all-link {
-		float: right;
-		&:after {
-			display: none;
-		}
-	}
-
-	.myft-feed {
-		padding: 0;
-		box-sizing: border-box;
-		margin: 0;
-	}
-
-	.myft-ui--prefer {
-		padding: 10px;
-	}
-
-
-	.myft-feed__item__tags {
-		display: none;
-	}
-
-	.myft-feed__item__headline {
-		@include nTypeAlpha(4);
-		margin: 0;
-
-		@include oGridRespondTo(S) {
-			@include nTypeAlpha(3);
-		}
-		> a {
-			color: oColorsGetPaletteColor('cold-2');
-			text-decoration: none;
-			&:hover {
-				text-decoration: underline;
-			}
-		}
-	}
-
-	.myft-feed__item {
-		list-style: none;
-		padding: 10px 0;
-		&:last-child {
-			border-bottom-width: 0;
-		}
-	}
-
-	.o-expander__collapsible-item {
-		position: static;
-		left: auto;
-	}
-
-	[aria-expanded="true"] .myft-feed__item {
-		display: list-item;
-		padding: 0;
-		border-bottom: 1px solid oColorsGetPaletteColor('warm-3');
-	}
-
-	&__title {
-		background: oColorsGetPaletteColor('red-1');
-		padding: 10px;
-		display: block;
-		margin: 0;
-		&,
-		a {
-			display: block;
-			color: oColorsGetPaletteColor('white');
-			text-transform: none;
-			text-decoration: none;
-			&:hover {
-				text-decoration: underline;
-			}
-		}
-
-		.myft-ui__icon {
-			margin: 0 4px -4px;
-		}
-	}
-
-	&__title-text {
-		@include nTypeBravo(4);
-		display: block;
-		padding: 0 0 6px;
-	}
-	&--from-email &__title-text:before {
-		@include nextIcon(email, oColorsGetPaletteColor('white'), 20);
-		content: '';
-		margin: -5px 0 0 -3px;
-		position: relative;
-		top: 5px;
-	}
-
-	&__next-headline {
-		@include nTypeAlpha(3);
-	}
-
-	&__current-page:before {
-		@include nTypeBravo(5);
-		content: 'Currently reading';
-		text-transform: uppercase;
-		display: block;
-	}
-
-
-	&__red-icon {
-		@include nextIcon(myft-bolder, oColorsGetPaletteColor('red-1'), 16);
-		width: 34px;
-		margin: 0 8px -12px;
-	}
-}
