+++ client/main.scss
@@ -50,6 +50,6 @@ $o-comments-is-silent: false;
 @import "components/comments/main";
 @import "components/first-click-free/main";
 @import "components/barriers/main";
-@import "components/reading-list/main";
+@import "components/email-signup/main";
 @import "components/typography-switcher/main";
 @import "components/special-report/main";
