+++ views/partials/email-signup.html
@@ -9,10 +9,10 @@
 			preferenceName="email-daily-digest"
 			label="Get a daily digest email with updates on the topics you follow" }}
 
-		{{> next-myft-ui/templates/prefer
-				buttonText="Subscribe"
-				alternateButtonText="Unsubscribe"
-				preferenceName="email-weekly-digest"
-				label="Get a weekly digest email with comment and analysis on the topics you follow" }}
+	{{> next-myft-ui/templates/prefer
+			buttonText="Subscribe"
+			alternateButtonText="Unsubscribe"
+			preferenceName="email-weekly-digest"
+			label="Get a weekly digest email with comment and analysis on the topics you follow" }}
 	</div>
 </div>
