--- views/partials/reading-list.html
@@ -1,13 +1,0 @@
-<div class="myft-reading-list js-myft-reading-list" data-trackable="myft-reading-list" data-o-grid-colspan="12 L8 XL7 XLoffset1" data-o-component="o-expander">
-	<a class="myft-reading-list__title" href="/myft/my-news" data-trackable="continue">
-		<span class="myft-reading-list__title-text">Continue reading your personal news feed at <span class="myft-ui__icon"><span class="u-visually-hidden">My FT</span></span></span>
-	</a>
-	<div class="more-on__content">
-
-		<div class="myft-reading-list__footer">
-			<a href="/myft/my-news" data-trackable="myft-icon"><span class="myft-reading-list__red-icon"><span class="u-visually-hidden">My FT</span></span></a>
-			<button class="more-on__all-link o-expander__toggle o--if-js" data-trackable="show-full-list"></button>
-		</div>
-	</div>
-</div>
-
